mod_http: Fix Basic Authentication decoding.
When receiving an Authorization header for Basic Authentication, both
the buffer and size passed to base64_decode were wrong. This manifested
as uninitialized memory accesses, and obviously it didn't work either.
The arguments are now correct, which takes care of the invalid memory
reads. A unit test has been added which captures both correctness and
the validity of memory accesses.

LBBS-122 #close