module.c: Fix module reference removal logic.
If there were multiple module refs from the same file with the same
pair ID (which is possible), the wrong module reference could be
removed, leading to possible use after free conditions. We now also
ensure that the the refmod pointer matches, which should be enough
to guarantee uniqueness and that the correct reference is removed.
Additional assertions have been added to catch this violation; these
previously failed but are no longer triggered with this fix.

LBBS-123 #close