net_imap: Allow direct proxied connections to remote servers.
Allow a user to login using <username>+<the proxy ID from .imapremote>
in order to establish an IMAP connection specifically to another
IMAP server, just using BBS credentials. This can be useful to
access another IMAP server as a dedicated account, especially if
the client does not support OAuth and needs the BBS to acts as
an OAuth proxy for authentication.