net_smtp: Allow SMTP hostname(s) to be set explicitly.
Hitherto, we have always use the BBS hostname as the SMTP hostname,
although we had a level of indirection in place to account for any
change here. Now, we allow this to be explicitly set by the user.
The incoming hostname used for the server banner on incoming connections
to us can also be set to yet a different hostname, if desired.
This is primarily useful if we want to the hostname used in bounces
generated by servers trying to deliver mail to us to be a different
hostname than the normal SMTP hostname, which should match with
reverse DNS and whose choice is thus somewhat restricted.