mod_asterisk_queues: Fix bug in agent status menu selection.
Commit 3890ba46fdba3f91723c85c07edf50ba6c1f6ddd added an option to
view menus showing agent statuses, and when agents are not a member
of all queues, the queue selection they have is more limited.
However, when choosing an option, we were not indexing the same way
as when creating the menu, where we skip queues in which the agent
is not a member. We now skip these when choosing the selection too.