nets/Makefile: Rebuild net_imap subdirectory object files if needed.
If header files in nets/net_imap were modified, net_imap.o would get
rebuilt (since it includes all of those header files), but the
corresponding object file in nets/net_imap would not get recompiled.
This can cause issues when make clean is not run between builds and
ABI changes, e.g. as a result of adding members in the middle of a
struct. To ensure these targets are rebuilt if needed, we now always
call nets/net_imap/Makefile as a dependency of the target for object
files in the subdirectory.

LBBS-114 #close