logger.c: Add support for automatic log rotation.
Allow the logger to be reloaded, which will simply reopen the log file.
This enables logrotate to be used, and a sample logrotate script is now
provided which can be installed using "make logrotate". The provided
script is mostly concerned with size, since the BBS log file can grow
to be very large with debug enabled.

Additionally, use useradd directly in scripts for compatibility, rather
than adduser, which is an abstraction on top of useradd.