net_imap: Add ability to hide SPECIAL-USE flags on proxied folders.
Add an experimental per-user setting to hide any SPECIAL-USE flags
on remote/proxied mailboxes. This setting was developed to see if
it would work around the bug where Mozilla-based mail clients will
treat any SPECIAL-USE folders in the "Other Users" namespace as
nearly identical siblings to the actual top-level SPECIAL-USE folder
for that account. This can cause issues, such as when a message
is deleted from the top-level account, and it goes to the Trash
folder in a proxied account instead of the account's actual Trash.

Unfortunately, Mozilla mail clients have enough logic to infer
SPECIAL-USE flags that simply hiding them is not sufficient in
all cases. However, this does help in the case where the remote
SPECIAL-USE folder names do not match a client's hardcoded list,
so even though this setting probably won't work very well in most
cases, I'm leaving it in for the time-being.