socket.c: Abort timed write during node/BBS shutdown.
timed_write has a 60-second timeout, which can delay shutdown
if a node is waiting to finish a write. Expedite this process
by polling internally for shorter periods of time and aborting
if a shutdown is active.

While not required, since the function will eventually
exit and shutdown eventually finalizes, this helps speed
up the process and make shutdown more responsive.