net_imap: Add ability to skip proxied LIST for certain clients.
The previously added 'disable_specialuse_by_agent' option in .imapremote
is not very effective; however, if we can hide all proxied folders
entirely in the LIST/LSUB response, then the client won't know about these
folders and thus can't make the mistakes that cause the issues discussed
in that option. While this will prevent use/access to these proxied
folders as part of accessing the user's BBS account via IMAP, if access
is not needed from particular clients (e.g. they are only accessed via
webmail, or through dedicated proxied connections to that account), then
this may not be an issue and it does reliably solve the issue.