mod_irc_relay: Allow AWAY to be mirrored on remote IRC networks.
For relays that are configured to operate as a private relay for a
specific IRC user, it may be desired to have remote IRC networks
reflect the status of the local user, such that if a user is marked
as AWAY locally, the status of the IRC client used for private
relays is also marked as AWAY. Add an option to allow for this.

As part of this addition, the IRC relay callbacks have been refactored
to make this and any future expansion less invasive, and a few other
minor (mostly formatting) issues have been fixed.