IRC: Automatically mark user away via relays at startup/join/part.
mod_irc_relay and mod_slack (the only two modules which currently
relay away/here status from IRC) now automatically mark the user
away via the relay if the IRC user is not online at the time that
the module loads (and vice versa, if the user is online).

Additionally, when an IRC user connects or disconnects, this is
also relayed as an away/back status as appropriate.

As a result, the user's status/presence on a remote IRC network or
on Slack can now mirror the user's local IRC status (away, if
set to away or not connected, and not away otherwise).

LBBS-27 #close