mod_slack: Don't require explicit mapping for receiving DMs.
Rearrange the logic in the on_message callback to allow any
DMs received appropriately addressed or associated to
notify the corresponding user. Previously, we would abort
if there was no mapping, but explicit mappings should not
be required for each individual Slack sender from whom we might
want to receive DMs.

Additionally, if delivery of a channel or private message via
a relay fails (Slack, notably), notify the sender on the IRC
side of this. Previously, a warning was logged to the console,
but the failure was not communicated to the actual user.