mod_slack: Avoid extraneously relaying messages containing URLs.
Normally, the message we received back from Slack is the same as
the one we posted; however, if it contains a URL, Slack surrounds
the URL with arrows (e.g. <https://example.com>). This would cause
us to echo the post back to IRC, even if it came from IRC. For
web URLs, this would even happen twice.

Now, if the message received from Slack is identical to our last
sent message, except for that it contains URLs surrounded by <>,
treat these as identical and don't relay it back.